# server/app/model/category.go
package model

import "time"

// Mapel merepresentasikan entitas Mata Pelajaran
type Category struct {
	ID        uint      `gorm:"primaryKey" json:"id"`
	Title     string    `json:"title"`
	Deskripsi string    `json:"deskripsi"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}